Finalize TaskListDAO

+++ src/main/java/lk/ijse/dep8/tasks/dao/UserDAO.java
@@ -1,5 +1,6 @@
 package lk.ijse.dep8.tasks.dao;
 
+import lk.ijse.dep8.tasks.dao.exception.DataAccessException;
 import lk.ijse.dep8.tasks.entity.User;
 
 import java.sql.*;
@@ -15,7 +16,7 @@ public class UserDAO {
         this.connection = connection;
     }
 
-    public boolean existsById(String userId) {
+    public boolean existsUserById(String userId) {
         try {
             PreparedStatement stm = connection.prepareStatement("SELECT id FROM user WHERE id=?");
             stm.setString(1, userId);
@@ -27,7 +28,7 @@ public class UserDAO {
 
     public User saveUser(User user) {
         try {
-            if (!existsById(user.getId())) {
+            if (!existsUserById(user.getId())) {
                 PreparedStatement stm = connection.
                         prepareStatement("INSERT INTO user (id, email, password, full_name, profile_pic) VALUES (?, ?, ?, ?, ?)");
                 stm.setString(1, user.getId());
@@ -58,6 +59,9 @@ public class UserDAO {
 
     public void deleteUserById(String userId) {
         try {
+            if (!existsUserById(userId)){
+                throw new DataAccessException("No user found");
+            }
             PreparedStatement stm = connection.prepareStatement("DELETE FROM user WHERE id=?");
             stm.setString(1, userId);
             if (stm.executeUpdate() != 1) {
